refactor: split alert ui components
